requirements.txt
fastapi
uvicorn
python-jose
requests
pydantic
slowapi
